# docs/swagger.yaml
basePath: /api/
definitions:
  apierr.ApiErr:
    properties:
      code:
        type: integer
      message: {}
    type: object
...
  dto.IdResponse:
    properties:
      id:
        example: 00000000-0000-0000-0000-000000000000
        type: string
      message:
        example: readme created successfully
        type: string
    required:
    - id
    type: object
...
  dto.SearchTemplateRequestDoc:
    properties:
      amount:
        minimum: 1
        type: integer
      filter:
        $ref: '#/definitions/dto.filterTemplatesFields'
      page:
        minimum: 1
        type: integer
      query:
        type: string
      sort:
        $ref: '#/definitions/dto.sortTemplatesFields'
    required:
    - amount
    - page
    type: object
...
  dto.SuccessResponse:
    properties:
      code:
        example: 200
        type: integer
      message:
        example: success
        type: string
    type: object
...
  dto.TemplateResponse:
    properties:
      description:
        minLength: 1
        type: string
      id:
        type: string
      image:
        type: string
      is_public:
        type: boolean
      last_update_time:
        type: string
      likes:
        minimum: 0
        type: integer
      num_of_users:
        minimum: 0
        type: integer
      owner_avatar:
        type: string
      owner_id:
        type: string
      owner_nickname:
        type: string
      title:
        minLength: 1
        type: string
    required:
    - description
    - id
    - image
    - is_public
    - last_update_time
    - likes
    - num_of_users
    - owner_avatar
    - owner_id
    - owner_nickname
    - title
    type: object
...
  dto.filterTemplatesFields:
    properties:
      isOfficial:
        type: boolean
    type: object
...
  dto.sortTemplatesFields:
    properties:
      LastUpdateTime:
        example: desc/asc
        type: string
      Likes:
        example: desc/asc
        type: string
      NumOfUsers:
        example: desc/asc
        type: string
    type: object
...
  models.TemplateWithOwner:
    properties:
      create_time:
        type: string
      description:
        type: string
      id:
        type: string
      image:
        type: string
      is_public:
        type: boolean
      last_update_time:
        type: string
      likes:
        type: integer
      links:
        items:
          type: string
        type: array
      num_of_users:
        type: integer
      owner_avatar:
        type: string
      owner_id:
        type: string
      owner_nickname:
        type: string
      render_order:
        items:
          type: string
        type: array
      text:
        items:
          type: string
        type: array
      title:
        type: string
      widgets:
        items:
          additionalProperties:
            type: string
          type: object
        type: array
    type: object
...
host: localhost:3333
info:
  contact:
    email: readmeow@gmail.com
    name: READMEOW Support
  description: Documentation to READMEOW API
  title: REODMEOW API
  version: 1.0.0
paths:
...
  /api/templates:
    get:
      consumes:
      - application/json
      description: Search templates with filters and sorting
      parameters:
      - description: Search templates request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SearchTemplateRequestDoc'
      produces:
      - application/json
      responses:
        "200":
          description: List of templates
          schema:
            items:
              $ref: '#/definitions/dto.TemplateResponse'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "422":
          description: Invalid JSON
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierr.ApiErr'
      summary: Search Templates
      tags:
      - Templates
    patch:
      description: Updating a user template
      parameters:
      - in: formData
        maxLength: 1000
        minLength: 1
        name: description
        type: string
      - in: formData
        name: id
        required: true
        type: string
      - format: binary
        in: formData
        name: image
        required: true
        type: string
      - in: formData
        name: is_public
        type: boolean
      - collectionFormat: csv
        in: formData
        items:
          type: string
        name: links
        type: array
      - collectionFormat: csv
        in: formData
        items:
          type: string
        name: render_order
        type: array
      - in: formData
        name: text
        type: string
      - in: formData
        maxLength: 255
        minLength: 1
        name: title
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "422":
          description: Invalid JSON
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierr.ApiErr'
      security:
      - ApiKeyAuth: []
      summary: Update Template
      tags:
      - Templates
    post:
      consumes:
      - multipart/form-data
      description: Creating a new template
      parameters:
      - in: formData
        maxLength: 1000
        minLength: 1
        name: description
        required: true
        type: string
      - format: binary
        in: formData
        name: image
        required: true
        type: string
      - in: formData
        name: is_public
        required: true
        type: boolean
      - collectionFormat: csv
        in: formData
        items:
          type: string
        name: links
        type: array
      - collectionFormat: csv
        in: formData
        items:
          type: string
        minItems: 1
        name: render_order
        required: true
        type: array
      - collectionFormat: csv
        in: formData
        items:
          type: string
        name: text
        type: array
      - in: formData
        maxLength: 255
        minLength: 1
        name: title
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            $ref: '#/definitions/dto.IdResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "409":
          description: Already exists
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "422":
          description: Invalid JSON
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierr.ApiErr'
      security:
      - ApiKeyAuth: []
      summary: Create Template
      tags:
      - Templates
  /api/templates/{template}:
    delete:
      description: Deleting a user template by its id
      parameters:
      - description: Template ID
        in: path
        name: template
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierr.ApiErr'
      security:
      - ApiKeyAuth: []
      summary: Delete Template
      tags:
      - Templates
    get:
      description: Get template by ID
      parameters:
      - description: Template ID
        in: path
        name: template
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Template data
          schema:
            $ref: '#/definitions/models.TemplateWithOwner'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierr.ApiErr'
      security:
      - ApiKeyAuth: []
      summary: Get Template
      tags:
      - Templates
  /api/templates/dislike/{template}:
    patch:
      description: Dislike template by ID
      parameters:
      - description: Template ID
        in: path
        name: template
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "409":
          description: Already disliked
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierr.ApiErr'
      security:
      - ApiKeyAuth: []
      summary: Dislike Template
      tags:
      - Templates
  /api/templates/favorite:
    get:
      consumes:
      - application/json
      description: Fetch favorite templates of current user
      parameters:
      - in: query
        minimum: 1
        name: amount
        required: true
        type: integer
      - in: query
        minimum: 1
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of favorite templates
          schema:
            items:
              $ref: '#/definitions/dto.TemplateResponse'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "422":
          description: Invalid JSON
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierr.ApiErr'
      security:
      - ApiKeyAuth: []
      summary: Fetch Favorite Templates
      tags:
      - Templates
  /api/templates/like/{template}:
    patch:
      description: Like template by ID
      parameters:
      - description: Template ID
        in: path
        name: template
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "409":
          description: Already liked
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierr.ApiErr'
      security:
      - ApiKeyAuth: []
      summary: Like Template
      tags:
      - Templates
...
swagger: "2.0"